ADDED new class activity folder
Code 2 script, one to get if a word it's palindrome or not and the second to set cap letters at the start of each word.

Additionally, this scripts should be done in two ways respectively, with loop and without loop (using recursion).

Also there are some other codes that my classmates Mabe and Ashley gave me as a "hint".