FINISHED the algorythm using RECURSION to verify is a word is palindrome or not.
First I used a recursive function to invert the phrase and then I was used a send function to verify if the
original phrase and the inverted phrase are the same, so while it's true for each character c = c + 1 the counter
to verify if the c (counter) and the Lenght of the phrase or inverted phrase are the same.
If so, it prints in terminal that the word is palindrome, else it's not!